requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distlib==0.3.6
Django==4.1.1
django-countries==7.3.2
filelock==3.8.0
idna==3.4
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.5
django-mptt~=0.14.0